object utility: "delete" to remove keys from bj [4:33pm]

+++ client/src/components/elements/LogInOutGoogle/LogInOutGoogle.tsx
@@ -30,8 +30,9 @@ import ConnectedSignupLoginChecker from '../../../components/elements/SignupLogi
 import SignupInput from '../../../components/elements/SignupInputs'
 
 import { connect, useDispatch } from 'react-redux'
-import { TOGGLE_LOGIN_SIGNUP_BTN, TOGGLE_SUBMIT_INPUT_DATA, TOGGLE_SHOW_FORM , SET_PASSWORD_INPUT, SET_ALL_USERS, SET_ALL_USERNAMES, SET_ALL_EMAILS, TOGGLE_GOOGLE_LINK_ACCT_SCREEN, SET_CURRENT_USER, TOGGLE_NO_LINK_GOOGLE_BTN_HOVER, TOGGLE_YES_LINK_GOOGLE_BTN_HOVER, TOGGLE_YES_LINK_GOOGLE_BTN_CLICK, TOGGLE_NO_LINK_GOOGLE_BTN_CLICK, SET_GOOGLE_IMG_URL, TOGGLE_ICON_NOT_INPUT, TOGGLE_PASSWORD_SHOW, TOGGLE_PASSWORD_SHOW_CLICK, SET_LOG_IN_OUT_FLASH_MSG, TOGGLE_USER_ICON_CONFIRM } from '../../../redux/actions'
+import { TOGGLE_LOGIN_SIGNUP_BTN, TOGGLE_SUBMIT_INPUT_DATA, TOGGLE_SHOW_FORM , SET_PASSWORD_INPUT, SET_ALL_USERS, SET_ALL_USERNAMES, SET_ALL_EMAILS, TOGGLE_GOOGLE_LINK_ACCT_SCREEN, SET_CURRENT_USER, TOGGLE_NO_LINK_GOOGLE_BTN_HOVER, TOGGLE_YES_LINK_GOOGLE_BTN_HOVER, TOGGLE_YES_LINK_GOOGLE_BTN_CLICK, TOGGLE_NO_LINK_GOOGLE_BTN_CLICK, SET_GOOGLE_IMG_URL, TOGGLE_ICON_NOT_INPUT, TOGGLE_PASSWORD_SHOW, TOGGLE_PASSWORD_SHOW_CLICK, SET_LOG_IN_OUT_FLASH_MSG, TOGGLE_USER_ICON_CONFIRM, SET_ONLINK_GOOGLE_CONFIRM_DATA } from '../../../redux/actions'
 import $ from 'jquery'
+import { Any } from "react-spring"
 // client/src/components/elements/LogInOutGoogle/LogInOutGoogle.module.scss // relative path for import above 
 
  function LogInOutGoogle ( props:any ) {
@@ -39,9 +40,9 @@ import $ from 'jquery'
     const { 
             LOGIN_SIGNUP_BTN, DISPLAY_FORM, INPUT_FOCUS, ALL_USERS, ALL_USERNAMES, USERNAME_INPUT, EMAIL_INPUT, PASSWORD_INPUT, AGE_INPUT, PARENT_CONFIRM, SUBMIT_INPUT_DATA, 
             TOGGLE_SUBMIT_INPUT_DATA, GOOGLE_LINK_ACCT_SCREEN, CURRENT_USER, NO_LINK_GOOGLE_BTN_HOVER,
-            YES_LINK_GOOGLE_BTN_HOVER, LINK_GOOGLE_BTN_CLICK, NO_LINK_GOOGLE_BTN_CLICK, GOOGLE_IMG_URL, ICON_NOT_INPUT, PASSWORD_SHOW, PASSWORD_SHOW_CLICK, LOG_IN_OUT_FLASH_MSG,
+            YES_LINK_GOOGLE_BTN_HOVER, LINK_GOOGLE_BTN_CLICK, NO_LINK_GOOGLE_BTN_CLICK, GOOGLE_IMG_URL, ICON_NOT_INPUT, PASSWORD_SHOW, PASSWORD_SHOW_CLICK, LOG_IN_OUT_FLASH_MSG, ONLINK_GOOGLE_CONFIRM_DATA,
                                     
-            TOGGLE_LOGIN_SIGNUP_BTN, TOGGLE_SHOW_FORM, SET_PASSWORD_INPUT, SET_ALL_USERS, SET_ALL_USERNAMES, TOGGLE_GOOGLE_LINK_ACCT_SCREEN, SET_CURRENT_USER,
+            TOGGLE_LOGIN_SIGNUP_BTN, TOGGLE_SHOW_FORM, SET_PASSWORD_INPUT, SET_ALL_USERS, SET_ALL_USERNAMES, TOGGLE_GOOGLE_LINK_ACCT_SCREEN, SET_CURRENT_USER, SET_ONLINK_GOOGLE_CONFIRM_DATA,
             TOGGLE_NO_LINK_GOOGLE_BTN_HOVER, TOGGLE_YES_LINK_GOOGLE_BTN_HOVER, TOGGLE_YES_LINK_GOOGLE_BTN_CLICK, TOGGLE_NO_LINK_GOOGLE_BTN_CLICK, SET_GOOGLE_IMG_URL, 
             TOGGLE_ICON_NOT_INPUT, TOGGLE_PASSWORD_SHOW, TOGGLE_PASSWORD_SHOW_CLICK, SET_LOG_IN_OUT_FLASH_MSG, TOGGLE_USER_ICON_CONFIRM,
           } = props
@@ -259,6 +260,7 @@ import $ from 'jquery'
             }
 
             const onLinkSuccess = async (res:any) => {
+
                 let RegexMenu = await RegexBank()
                 let noWhiteSpaceRegex = RegexMenu.noWhiteSpace
                                         
@@ -270,10 +272,17 @@ import $ from 'jquery'
         
                 deathCertificate('googleImgUrl', googleImgUrl, 1, false)
                 SET_GOOGLE_IMG_URL({ payload: googleImgUrl })
-
                 
-                let userUpdatedWithGoogle = await linkUserWithGoogle(USERNAME_INPUT, googleId, googleImgUrl)                                
+                let userUpdatedWithGoogle = await linkUserWithGoogle(USERNAME_INPUT, googleId, googleImgUrl)      
                 const u = userUpdatedWithGoogle.data.linkUserWithGoogle
+                console.log('u')                
+                console.log(u)
+
+                console.log('ONLINK_GOOGLE_CONFIRM_DATA')
+                console.log(ONLINK_GOOGLE_CONFIRM_DATA)
+                // SET_ONLINK_GOOGLE_CONFIRM_DATA( {payload: { id: u.id, username: u.username, icon: u.icon, googleId: u.googleId, age: u.age}})                   
+
+
 
                 const uGID = u.googleId
                 if (noWhiteSpaceRegex.test(uGID)) {
@@ -282,20 +291,16 @@ import $ from 'jquery'
                     setTimeout( () => {
                     SET_LOG_IN_OUT_FLASH_MSG( { payload: '' })
                     }, 6000)
-                } 
-                                
-                console.log('userUpdatedWithGoogle')
-                console.log(userUpdatedWithGoogle)
-            
+                }                                         
                 let userStringObject = `GID:${u.googleId},icon:${u.icon},username:${u.username},password:${u.password},age:${u.age},id:${u.id}`
 
                 await localStorage.setItem("user", userStringObject)
 
-                timer = await setTimeout(async() => {
-                    let updatedUserStrObj:any = await localStorage.getItem("user");
-                    console.log('updatedUserStrObj')
-                    console.log(updatedUserStrObj)
-                }, 2000)
+                // timer = await setTimeout(async() => {
+                //     let updatedUserStrObj:any = await localStorage.getItem("user");
+                //     console.log('updatedUserStrObj')
+                //     console.log(updatedUserStrObj)
+                // }, 2000)
 
             }
         
@@ -307,28 +312,41 @@ import $ from 'jquery'
                 })
                 
                 urlPROMISE.then( (urldata:any) => {
-                    let localNODE_ENV = urldata.ENVdata.data.ENV.NODE_ENV                    
+                    let localNODE_ENV = urldata.ENVdata .data.ENV.NODE_ENV                    
                     const saveUserPROMISE = new Promise( (resolve, reject) => {
-                        let userSignupURL = urldata.userSignupURL
+                        
                         resolve(userSignup({ googleId: '', icon: '', username: USERNAME_INPUT, email: EMAIL_INPUT, password: PASSWORD_INPUT, age: AGE_INPUT }, localNODE_ENV))
+                        // resolve(userSignup({ googleId: '', icon: '', username: USERNAME_INPUT, email: EMAIL_INPUT, password: PASSWORD_INPUT, age: AGE_INPUT }, localNODE_ENV))
                         reject([])
                     })
-                    saveUserPROMISE.then(async(userSignup:any) => {         
-
-                        console.log('userSignup')            
-                        console.log(userSignup)            
-
-                                                
-
-                        let userStringForLocStorage = JSON.stringify(userSignup)
-                        localStorage.setItem('wateruser', userStringForLocStorage)
-                        localStorage.setItem('GTOKEN', "GOOGLE")
-                        console.log('userStringForLocStorage')
-                        console.log(userStringForLocStorage)
+                    saveUserPROMISE.then(async(userSignup:any) => {  
+                        console.log('userSignup this is easy')
+                        console.log(userSignup)
                         
+                        const  toggle_link_Promise = new Promise(async(resolve:any, reject:any) => {
+                            let u = userSignup
+                            let newU = {}
+                            await delete u.data.userSignup.password      
+                            console.log('u after the key is altered')                                                                                                          
+                            console.log(u)                                                                                                          
+                            await SET_ONLINK_GOOGLE_CONFIRM_DATA({payload: u })                        
+                            let confirmtoken = ONLINK_GOOGLE_CONFIRM_DATA.length > 1 ? ONLINK_GOOGLE_CONFIRM_DATA : "reject"
+                            resolve(confirmtoken)
+                            // reject(confirmtoken) 
+                        }).then( (noPWuser:any) => {
+                            console.log('noPWuser')
+                            console.log(noPWuser)
+                            TOGGLE_YES_LINK_GOOGLE_BTN_CLICK()
+                            localStorage.setItem('GTOKEN', "GOOGLE")
+                        })
 
+
+                        // let userStringForLocStorage = JSON.stringify(userSignup)
+                        // localStorage.setItem('wateruser', userStringForLocStorage)
+
+                        
+                
                         // UPDATE TO        G       O       O       G       L       E               !!!!!!!!!!!!!!!!!!!
-                        TOGGLE_YES_LINK_GOOGLE_BTN_CLICK()
 
                         // setTimeout( () => {
                         //     lazyJQanimate($('#link-google'), 'link-google', 'signup')     
@@ -548,6 +566,7 @@ const mapStateToProps = (state:any) => ({
     NO_LINK_GOOGLE_BTN_CLICK: state.NO_LINK_GOOGLE_CLICK,
     GOOGLE_IMG_URL: state.GOOGLE_IMG_URL,
     ICON_NOT_INPUT: state.ICON_NOT_INPUT,
+    ONLINK_GOOGLE_CONFIRM: state.ONLINK_GOOGLE_CONFIRM
 })
 
 const mapDispatchToProps = (dispatch:any) => ({
@@ -570,7 +589,8 @@ const mapDispatchToProps = (dispatch:any) => ({
     TOGGLE_NO_LINK_GOOGLE_BTN_CLICK: () => dispatch(TOGGLE_NO_LINK_GOOGLE_BTN_CLICK()),
     SET_GOOGLE_IMG_URL: (action:any) => dispatch(SET_GOOGLE_IMG_URL(action)),
     TOGGLE_ICON_NOT_INPUT: () => dispatch(TOGGLE_ICON_NOT_INPUT()),
-    TOGGLE_USER_ICON_CONFIRM: () => dispatch(TOGGLE_USER_ICON_CONFIRM())
+    TOGGLE_USER_ICON_CONFIRM: () => dispatch(TOGGLE_USER_ICON_CONFIRM()),
+    SET_ONLINK_GOOGLE_CONFIRM_DATA: (action:any) => dispatch(SET_ONLINK_GOOGLE_CONFIRM_DATA(action))
     // TOGGLE_HYDRO_SETTINGS: () => dispatch(TOGGLE_HYDRO_SETTINGS()),
 })
